+++ NEWS.md
@@ -1,5 +1,9 @@
 # supernova 2.3.0
 
+* Dependency on `lme4` is moved to Suggests. Models implementing `lmerMod` are handled via `supernova.lmerMod` and `variables.lmerMod` but use of the `lme4` package is limited to tests.
+* More robust and readable implementation of `variables()` using the new formula utility functions added. See `?formula_building`, `?formula_expansion`, and `?formula_extraction`.
+
+
 # supernova 2.2.3
 
 * Remove dependency on `dplyr` because it changes too quickly and has too many other dependencies