+++ NEWS.md
@@ -2,7 +2,7 @@
 
 * Dependency on `lme4` is moved to Suggests. Models implementing `lmerMod` are handled via `supernova.lmerMod` and `variables.lmerMod` but use of the `lme4` package is limited to tests.
 * More robust and readable implementation of `variables()` using the new formula utility functions added. See `?formula_building`, `?formula_expansion`, and `?formula_extraction`.
-* Customize `print.lm()` to add the fitted equation at the end of the printed output (thanks for the suggestion from [@ave-63](https://github.com/ave-63)!)
+* Add a new function `equation()` to extract the fitted equation from a linear model (`lm()`) (thanks for the suggestion from [@ave-63](https://github.com/ave-63)!)
 
 
 # supernova 2.2.3