+++ NEWS.md
@@ -1,4 +1,4 @@
-# Pre-release
+# supernova 2.4.0
 
 ## New features